# [scheduler] Various optimizations to task stealing