Profiling some services on many-core machines showed very high lock contention in the scheduler, specifically in the task-stealing mechanism.
This PR has a bunch of commits that both shrink the related critical sections, and reduce contention on these locks.

Testing:
- make emulator_test ARGS="-suite process_SUITE"
- tested in production on a many-core server, it showed very large improvements at low to medium load (on the order of 30% vs 40% cpu utilization), and tiny but measurable improvements at high load (about 1-2% cpu utilization). Lock-counting confirmed a fall in the contention of the runqueue locks.