@garazdawi 2025-03-20
> I could not reproduce the failure of one test in monitor_SUITE locally.  And is this test known to be flaky?

We have a couple of failures of that testcase in our lab, so yes it can be a bit flaky. I re-triggered the tests so that we can see if it fails again.