@garazdawi 2025-03-20
I restarted the CI jobs again in order for this PR to run the tests in all applications, not just emulator.