@garazdawi 2025-03-20
You can have a look here to see which tests currently fail on a clean master build: https://erlang.org/github-pr/master/ct_logs/