@garazdawi 2025-03-20
> I restarted the CI jobs again in order for this PR to run the tests in all applications, not just emulator.

apparently that did not work, it should work next time you push a fix to the PR though.