@tomas-abrahamsson 2025-03-26
This looks very interesting. I took the [8ac8af9f82 Silence (spurious) compiler warning](https://github.com/erlang/otp/pull/9594/commits/8ac8af9f82701913f04eb280f7905d7a5ca82fd2) commit for a run on one of our servers, using the commit that this branch is based on, the [57cee85703](https://github.com/erlang/otp/commit/57cee857031cc08a47c7d2f4f21a6fac87781f91) as baseline for a comparison. I captured some lock-count profilings, don't know if this is interesting to anyone, but anyways, here goes.

The baseline 28.0 rc was already about 11% faster than 27.3 for our workload, very nice, and this PR adds another 11% Very very nice! :)  For the `run_queue` lock:
* 'time [us]' is reduced by a factor 10.5
* '#collisions' is reduced by a factor 4.5
* '#tries' is reduced by a factor 3.2

Baseline commit [57cee85703](https://github.com/erlang/otp/commit/57cee857031cc08a47c7d2f4f21a6fac87781f91), the first few lines:
```
(csim-gte-2363519-19560@somehost3003)1> lcnt:start(),
                                        lcnt:collect(),
                                        timer:sleep(3_000),
                                        lcnt:conflicts(),
                                        lcnt:stop().
Command is taking a long time, type Ctrl+G, then enter 'i' to interrupt
                       lock     id    #tries  #collisions  collisions [%]  time [us]  duration [%]
                      -----    ---   ------- ------------ --------------- ---------- -------------
                  run_queue     82 377823485     38788195         10.2662   91615112       32.0361
                  proc_main 247945 106692940      5139515          4.8171    4230013        1.4792
                   pix_lock   1024     28682           66          0.2301    2756360        0.9638
        export_staging_lock      1    777325           37          0.0048    1320580        0.4618
                  proc_msgq 247945 178526735       727267          0.4074     946128        0.3308
```

The [8ac8af9f82 Silence (spurious) compiler warning](https://github.com/erlang/otp/pull/9594/commits/8ac8af9f82701913f04eb280f7905d7a5ca82fd2) commit:
```
                       lock     id    #tries  #collisions  collisions [%]  time [us]  duration [%]
                      -----    ---   ------- ------------ --------------- ---------- -------------
                  run_queue     82 118712446      8677443          7.3096    8749167        4.9924
                  proc_main 248331  38812418      1889950          4.8694    2984591        1.7030
                   pix_lock   1024     28195           71          0.2518    2679418        1.5289
        export_staging_lock      1    447291           33          0.0074    1023174        0.5838
                  proc_msgq 248331  63679742       309627          0.4862     824230        0.4703
```
The workload is a mixture of communication: a lot of Erlang processes, around 250k, that communicate both via direct message passing, and also via many TCP and SCTP connections. The hardware in this case was a server with 2×Xeon Gold 6148 with 20 cores each, in total 80 logical CPUs (2 threads per core). We have some newer machines too with more cores, both dual and single socket servers, but I haven't checked those.