@RobinMorisset 2025-03-27
> This looks very interesting. I took the [8ac8af9f82 Silence (spurious) compiler warning](https://github.com/erlang/otp/pull/9594/commits/8ac8af9f82701913f04eb280f7905d7a5ca82fd2) commit for a run on one of our servers, using the commit that this branch is based on, the [57cee85703](https://github.com/erlang/otp/commit/57cee857031cc08a47c7d2f4f21a6fac87781f91) as baseline for a comparison. I captured some lock-count profilings, don't know if this is interesting to anyone, but anyways, here goes.
> 
> The baseline 28.0 rc was already about 11% faster than 27.3 for our workload, very nice, and this PR adds another 11% Very very nice! :) For the `run_queue` lock:

Thanks a lot for these detailed numbers, they are very encouraging! Do you remember roughly what the CPU utilisation was when you took these measurements? We sadly observed smaller wins when the CPU is above 90% (since there is much less task stealing going on, this is not very surprising).