@tomas-abrahamsson 2025-03-29
Ok that's useful to know, and makes sense. The CPU load varied over the logical CPUs. For maybe 55-60 of the 80, it was around 80-90%, while for the remaining it was lower, much lower for a few cases. (I think we have some more bottlenecks to find in our code.)